+++ tests/Integration/ScriptsTest.php
@@ -402,50 +402,4 @@ final class ScriptsTest extends TestCase {
 			);
 		}
 	}
-
-	/**
-	 * Test if the web stories analytics got enqueued.
-	 *
-	 * @since 3.2.0
-	 *
-	 * @covers \Parsely\Scripts::run()
-	 * @group scripts
-	 */
-	public function test_web_stories_script_is_enqueued(): void {
-		self::$scripts->run();
-		self::assertSame(
-			10,
-			has_action( 'web_stories_print_analytics', array( self::$scripts, 'render_amp_analytics_tracker' ) )
-		);
-	}
-
-	/**
-	 * Test if the AMP tracker render outputs the correct script.
-	 *
-	 * @since 3.2.0
-	 *
-	 * @covers \Parsely\Scripts::render_amp_analytics_tracker
-	 * @group scripts
-	 */
-	public function test_render_amp_analytics_tracker(): void {
-		ob_start();
-		$this::$scripts->render_amp_analytics_tracker();
-		$output = ob_get_clean();
-		$output = str_replace( array( "\r", "\n" ), '', $output );
-
-		$expected = '
-		<amp-analytics type="parsely">
-			<script type="application/json">
-				{
-					"vars": {
-						"apikey": "blog.parsely.com"
-					}
-				}
-			</script>
-		</amp-analytics>
-		';
-		$expected = str_replace( array( "\r", "\n" ), '', $expected );
-
-		self::assertSame( $expected, $output );
-	}
 }
